Add enums and fix model type issues
Add LeaseStatus, MaintenanceStatus, UrgencyLevel and PaymentMethod enums. Fix ApplicationUser to User in MaintenanceRequest and FloorArea type from double to decimal in Unit.

+++ PropertyManagement/Models/Unit.cs
@@ -15,7 +15,7 @@ public class Unit
     [MaxLength(100)]
     public string Type { get; set; } = string.Empty;
 
-    public double FloorArea { get; set; }
+    public decimal FloorArea { get; set; }
 
     public int Bedrooms { get; set; }
